requirements.txt
streamlit==1.27.0
st_pages
numpy
pandas
matplotlib>=3.6.2
seaborn>=0.12.0
scikit_learn>=1.1.0
plotly>=5.11.0
streamlit-toggle-switch
-e .